api/requirements.txt
Django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
Pillow
django-filter
psycopg2-binary